Move GitHub issue to correct package and fix problem with multiple @Test annotations
