+++ src/test/java/spoon/reflect/factory/PackageFactoryTest.java
@@ -3,13 +3,14 @@ package spoon.reflect.factory;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.sameInstance;
 import static org.hamcrest.MatcherAssert.assertThat;
-
+import org.junit.jupiter.api.Test;
 import spoon.Launcher;
 import spoon.reflect.declaration.CtPackage;
-import spoon.test.GitHubIssue;
+import spoon.testing.utils.GitHubIssue;
 
 class PackageFactoryTest {
 
+	@Test
 	@GitHubIssue(issueNumber = 4764, fixed = true)
 	void getOrCreate_returnsNestedPackageStructure_whenQualifiedNameRepeatsSimpleName() {
 		// contract: A qualified name that is simply repetitions of a single simple name results in the expected
