+++ src/test/java/spoon/support/visitor/java/JavaReflectionTreeBuilderTest.java
@@ -106,13 +106,13 @@ import spoon.support.reflect.declaration.CtFieldImpl;
 import spoon.support.util.compilation.JavacFacade;
 import spoon.support.visitor.equals.EqualsChecker;
 import spoon.support.visitor.equals.EqualsVisitor;
-import spoon.test.GitHubIssue;
 import spoon.test.generics.testclasses3.ComparableComparatorBug;
 import spoon.test.innerclasses.InnerClasses;
 import spoon.test.pkg.PackageTest;
 import spoon.test.pkg.cyclic.Outside;
 import spoon.test.pkg.cyclic.direct.Cyclic;
 import spoon.test.pkg.cyclic.indirect.Indirect;
+import spoon.testing.utils.GitHubIssue;
 
 public class JavaReflectionTreeBuilderTest {
 
@@ -850,6 +850,7 @@ public class JavaReflectionTreeBuilderTest {
 		assertThat(asClass.getConstructors().iterator().next().getParameters().size(), equalTo(inners.size()));
 	}
 
+	@Test
 	@GitHubIssue(issueNumber = 4972, fixed = true)
 	void parameterNamesAreParsedWhenCompilingWithParametersFlag() throws ClassNotFoundException {
 		ClassLoader loader = JavacFacade.compileFiles(
